# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeOS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Memory {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  date: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  semester?: number | null;
  tags?: string[];
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface MemoryInput {
  title: string;
  description?: string;
  date: string;
  location?: string;
  semester?: number;
  tags?: string[];
  imageUrl?: string;
}

export interface MemoryUpdate {
  title?: string;
  description?: string;
  date?: string;
  location?: string;
  semester?: number;
  tags?: string[];
  imageUrl?: string;
}

export interface Note {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  url?: string | null;
  group: string;
  tags?: string[];
  createdAt: string;
}

export interface NoteInput {
  title: string;
  description?: string;
  url?: string;
  group: string;
  tags?: string[];
}

export interface NoteUpdate {
  title?: string;
  description?: string;
  url?: string;
  group?: string;
  tags?: string[];
}

export interface Subject {
  id: number;
  name: string;
  /** @nullable */
  code?: string | null;
  semester: number;
  credits: number;
  /** @nullable */
  gradePoint?: number | null;
  /** @nullable */
  grade?: string | null;
}

export interface SubjectInput {
  name: string;
  code?: string;
  semester: number;
  credits: number;
  gradePoint?: number;
  grade?: string;
}

export interface SubjectUpdate {
  name?: string;
  code?: string;
  semester?: number;
  credits?: number;
  gradePoint?: number;
  grade?: string;
}

export interface Assignment {
  id: number;
  title: string;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  dueDate: string;
  priority?: string;
  status: string;
  /** @nullable */
  semester?: number | null;
  createdAt: string;
}

export interface AssignmentInput {
  title: string;
  subjectId?: number;
  dueDate: string;
  priority?: string;
  status?: string;
  semester?: number;
}

export interface AssignmentUpdate {
  title?: string;
  subjectId?: number;
  dueDate?: string;
  priority?: string;
  status?: string;
  semester?: number;
}

export interface Exam {
  id: number;
  title: string;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  examDate: string;
  category: string;
  /** @nullable */
  semester?: number | null;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  gradePoint?: number | null;
  createdAt: string;
}

export interface ExamInput {
  title: string;
  subjectId?: number;
  examDate: string;
  category: string;
  semester?: number;
  grade?: string;
  gradePoint?: number;
}

export interface ExamUpdate {
  title?: string;
  subjectId?: number;
  examDate?: string;
  category?: string;
  semester?: number;
  grade?: string;
  gradePoint?: number;
}

export interface Achievement {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  date: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  category?: string | null;
  createdAt: string;
}

export interface AchievementInput {
  title: string;
  description?: string;
  date: string;
  icon?: string;
  category?: string;
}

export interface AchievementUpdate {
  title?: string;
  description?: string;
  date?: string;
  icon?: string;
  category?: string;
}

export interface Project {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  liveUrl?: string | null;
  tags?: string[];
  /** @nullable */
  semester?: number | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface ProjectInput {
  title: string;
  description?: string;
  githubUrl?: string;
  liveUrl?: string;
  tags?: string[];
  semester?: number;
  status?: string;
  imageUrl?: string;
}

export interface ProjectUpdate {
  title?: string;
  description?: string;
  githubUrl?: string;
  liveUrl?: string;
  tags?: string[];
  semester?: number;
  status?: string;
  imageUrl?: string;
}

export interface BucketListItem {
  id: number;
  text: string;
  checked: boolean;
  /** @nullable */
  targetDate?: string | null;
  createdAt: string;
}

export interface BucketListItemInput {
  text: string;
  targetDate?: string;
}

export interface BucketListItemUpdate {
  text?: string;
  checked?: boolean;
  targetDate?: string;
}

export interface Friend {
  id: number;
  name: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  birthday?: string | null;
  /** @nullable */
  firstMet?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  createdAt: string;
}

export interface FriendInput {
  name: string;
  department?: string;
  birthday?: string;
  firstMet?: string;
  notes?: string;
  photoUrl?: string;
}

export interface FriendUpdate {
  name?: string;
  department?: string;
  birthday?: string;
  firstMet?: string;
  notes?: string;
  photoUrl?: string;
}

export interface JournalEntry {
  id: number;
  date: string;
  /** @nullable */
  mood?: string | null;
  /** @nullable */
  bestMoment?: string | null;
  /** @nullable */
  learned?: string | null;
  /** @nullable */
  watched?: string | null;
  /** @nullable */
  randomThought?: string | null;
  createdAt: string;
}

export interface JournalEntryInput {
  date: string;
  mood?: string;
  bestMoment?: string;
  learned?: string;
  watched?: string;
  randomThought?: string;
}

export interface JournalEntryUpdate {
  mood?: string;
  bestMoment?: string;
  learned?: string;
  watched?: string;
  randomThought?: string;
}

export interface MoodLog {
  id: number;
  date: string;
  mood: string;
}

export interface MoodLogInput {
  mood: string;
  date?: string;
}

export interface FutureLetter {
  id: number;
  title: string;
  /** @nullable */
  content?: string | null;
  unlockDate: string;
  isUnlocked?: boolean;
  createdAt: string;
}

export interface FutureLetterInput {
  title: string;
  content: string;
  unlockDate: string;
}

export interface Stats {
  totalMemories: number;
  totalNotes: number;
  totalProjects: number;
  totalAchievements: number;
  totalFriends: number;
  totalDaysRecorded: number;
  assignmentsCompleted: number;
  assignmentsPending: number;
  bucketListTotal: number;
  bucketListCompleted: number;
}

export interface GpaSemesterData {
  semester: number;
  gpa: number;
  cgpa: number;
  credits?: number;
}

export type ListMemoriesParams = {
/**
 * @nullable
 */
semester?: number | null;
/**
 * @nullable
 */
tag?: string | null;
/**
 * @nullable
 */
year?: number | null;
/**
 * @nullable
 */
search?: string | null;
};

export type GetRecentMemoriesParams = {
limit?: number;
};

export type ListNotesParams = {
/**
 * @nullable
 */
group?: string | null;
/**
 * @nullable
 */
tag?: string | null;
/**
 * @nullable
 */
search?: string | null;
};

export type GetRecentNotesParams = {
limit?: number;
};

export type ListSubjectsParams = {
/**
 * @nullable
 */
semester?: number | null;
};

export type ListAssignmentsParams = {
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
semester?: number | null;
};

export type GetUpcomingAssignmentsParams = {
limit?: number;
};

export type ListExamsParams = {
/**
 * @nullable
 */
semester?: number | null;
};

export type ListMoodLogsParams = {
/**
 * @nullable
 */
range?: string | null;
};