# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LifeOS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memories:
    get:
      operationId: listMemories
      tags: [memories]
      summary: List all memories
      parameters:
        - in: query
          name: semester
          schema: { type: ["integer", "null"] }
        - in: query
          name: tag
          schema: { type: ["string", "null"] }
        - in: query
          name: year
          schema: { type: ["integer", "null"] }
        - in: query
          name: search
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: List of memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
    post:
      operationId: createMemory
      tags: [memories]
      summary: Create a memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
...
  /memories/recent:
    get:
      operationId: getRecentMemories
      tags: [memories]
      summary: Get recent memories
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 3 }
      responses:
        "200":
          description: Recent memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
  /memories/{id}:
    get:
      operationId: getMemory
      tags: [memories]
      summary: Get a memory by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "404":
          description: Not found
    patch:
      operationId: updateMemory
      tags: [memories]
      summary: Update a memory
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
    delete:
      operationId: deleteMemory
      tags: [memories]
      summary: Delete a memory
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # NOTES
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all notes
      parameters:
        - in: query
          name: group
          schema: { type: ["string", "null"] }
        - in: query
          name: tag
          schema: { type: ["string", "null"] }
        - in: query
          name: search
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/recent:
    get:
      operationId: getRecentNotes
      tags: [notes]
      summary: Get recent notes
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 3 }
      responses:
        "200":
          description: Recent notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # SUBJECTS
...
  /journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List journal entries
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Create or upsert today's journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
  /journal/today:
    get:
      operationId: getTodayJournal
      tags: [journal]
      summary: Get today's journal entry
      responses:
        "200":
          description: Today's journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          description: No entry today
  /journal/{id}:
    patch:
      operationId: updateJournalEntry
      tags: [journal]
      summary: Update a journal entry
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"

  # MOOD
  /mood:
    get:
      operationId: listMoodLogs
      tags: [mood]
      summary: List mood log entries
      parameters:
        - in: query
          name: range
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: Mood logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MoodLog"
    post:
      operationId: logMood
      tags: [mood]
      summary: Log mood for today (upsert)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoodLogInput"
      responses:
        "200":
          description: Mood logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodLog"

  # FUTURE LETTERS
...
components:
  schemas:
...
    Memory:
      type: object
      required: [id, title, date, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        date: { type: string }
        location: { type: ["string", "null"] }
        semester: { type: ["integer", "null"] }
        tags: { type: array, items: { type: string } }
        imageUrl: { type: ["string", "null"] }
        createdAt: { type: string }
    MemoryInput:
      type: object
      required: [title, date]
      properties:
        title: { type: string }
        description: { type: string }
        date: { type: string }
        location: { type: string }
        semester: { type: integer }
        tags: { type: array, items: { type: string } }
        imageUrl: { type: string }
    MemoryUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        date: { type: string }
        location: { type: string }
        semester: { type: integer }
        tags: { type: array, items: { type: string } }
        imageUrl: { type: string }
    Note:
      type: object
      required: [id, title, group, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        url: { type: ["string", "null"] }
        group: { type: string }
        tags: { type: array, items: { type: string } }
        createdAt: { type: string }
    NoteInput:
      type: object
      required: [title, group]
      properties:
        title: { type: string }
        description: { type: string }
        url: { type: string }
        group: { type: string }
        tags: { type: array, items: { type: string } }
    NoteUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        url: { type: string }
        group: { type: string }
        tags: { type: array, items: { type: string } }
...
    JournalEntry:
      type: object
      required: [id, date, createdAt]
      properties:
        id: { type: integer }
        date: { type: string }
        mood: { type: ["string", "null"] }
        bestMoment: { type: ["string", "null"] }
        learned: { type: ["string", "null"] }
        watched: { type: ["string", "null"] }
        randomThought: { type: ["string", "null"] }
        createdAt: { type: string }
    JournalEntryInput:
      type: object
      required: [date]
      properties:
        date: { type: string }
        mood: { type: string }
        bestMoment: { type: string }
        learned: { type: string }
        watched: { type: string }
        randomThought: { type: string }
    JournalEntryUpdate:
      type: object
      properties:
        mood: { type: string }
        bestMoment: { type: string }
        learned: { type: string }
        watched: { type: string }
        randomThought: { type: string }
    MoodLog:
      type: object
      required: [id, date, mood]
      properties:
        id: { type: integer }
        date: { type: string }
        mood: { type: string }
    MoodLogInput:
      type: object
      required: [mood]
      properties:
        mood: { type: string }
        date: { type: string }
...