# lib/api-zod/src/generated/types/journalEntryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeOS API
 * OpenAPI spec version: 0.1.0
 */

export interface JournalEntryUpdate {
  mood?: string;
  bestMoment?: string;
  learned?: string;
  watched?: string;
  randomThought?: string;
}